// assets/js/data/payment-methods/action-types.ts
export enum ACTION_TYPES {
	SET_PAYMENT_METHODS_INITIALIZED = 'SET_PAYMENT_METHODS_INITIALIZED',
	SET_EXPRESS_PAYMENT_METHODS_INITIALIZED = 'SET_EXPRESS_PAYMENT_METHODS_INITIALIZED',
	SET_ACTIVE_PAYMENT_METHOD = 'SET_ACTIVE_PAYMENT_METHOD',
	SET_SHOULD_SAVE_PAYMENT_METHOD = 'SET_SHOULD_SAVE_PAYMENT_METHOD',
	SET_AVAILABLE_PAYMENT_METHODS = 'SET_AVAILABLE_PAYMENT_METHODS',
	SET_AVAILABLE_EXPRESS_PAYMENT_METHODS = 'SET_AVAILABLE_EXPRESS_PAYMENT_METHODS',
	SET_PAYMENT_STATUS = 'SET_PAYMENT_STATUS',
	ADD_REGISTERED_PAYMENT_METHOD = 'ADD_REGISTERED_PAYMENT_METHOD',
	ADD_REGISTERED_EXPRESS_PAYMENT_METHOD = 'ADD_REGISTERED_EXPRESS_PAYMENT_METHOD',
	INITIALIZE_PAYMENT_METHODS = 'INITIALIZE_PAYMENT_METHODS',
	REMOVE_REGISTERED_PAYMENT_METHOD = 'REMOVE_REGISTERED_PAYMENT_METHOD',
	SET_PAYMENT_METHOD_DATA = 'SET_PAYMENT_METHOD_DATA',
	REMOVE_REGISTERED_EXPRESS_PAYMENT_METHOD = 'REMOVE_REGISTERED_EXPRESS_PAYMENT_METHOD',
}